TPIII/Hit3/worker/requirements.txt
fastapi
uvicorn[standard]
pika
redis
pillow
numpy
scipy